move semaphore fields
